+++ VisInfProc/cuda_filtermanager.cu
@@ -5,8 +5,8 @@
 
 
 __global__ void kernelCombineFilters(int sx, int sy, int sz,
-                                     double* gpuTemp, double* gpuSpatial,
-                                     double* gpuCombined){
+                                     float* gpuTemp, float* gpuSpatial,
+                                     float* gpuCombined){
     int combinedIdx = threadIdx.x + blockIdx.x * blockDim.x;
     int s_xy = sx*sy;
     int z = combinedIdx / s_xy;
@@ -16,8 +16,8 @@ __global__ void kernelCombineFilters(int sx, int sy, int sz,
 }
 
 __host__ void cudaCombineFilters(int sx, int sy, int sz,
-                                 double* gpuTemp, double* gpuSpatial,
-                                 double* gpuCombined)
+                                 float* gpuTemp, float* gpuSpatial,
+                                 float* gpuCombined)
 {
 
     long blocks = ceil((float)(sx*sy*sz)/THREADS_PER_BLOCK);
